+++ src/RcppExports.cpp
@@ -50,6 +50,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// avgrange
+pADFun avgrange(pADFun& pfun);
+RcppExport SEXP _scorematchingad_avgrange(SEXP pfunSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< pADFun& >::type pfun(pfunSEXP);
+    rcpp_result_gen = Rcpp::wrap(avgrange(pfun));
+    return rcpp_result_gen;
+END_RCPP
+}
 // set_cppad_error_handler
 void set_cppad_error_handler();
 RcppExport SEXP _scorematchingad_set_cppad_error_handler() {
@@ -69,15 +80,15 @@ BEGIN_RCPP
 END_RCPP
 }
 // fixdynamic
-pADFun fixdynamic(pADFun& uld, veca1 theta, Eigen::Matrix<int, Eigen::Dynamic, 1> fixedtheta);
-RcppExport SEXP _scorematchingad_fixdynamic(SEXP uldSEXP, SEXP thetaSEXP, SEXP fixedthetaSEXP) {
+pADFun fixdynamic(pADFun& pfun, veca1 theta, Eigen::Matrix<int, Eigen::Dynamic, 1> isfixed);
+RcppExport SEXP _scorematchingad_fixdynamic(SEXP pfunSEXP, SEXP thetaSEXP, SEXP isfixedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< pADFun& >::type uld(uldSEXP);
+    Rcpp::traits::input_parameter< pADFun& >::type pfun(pfunSEXP);
     Rcpp::traits::input_parameter< veca1 >::type theta(thetaSEXP);
-    Rcpp::traits::input_parameter< Eigen::Matrix<int, Eigen::Dynamic, 1> >::type fixedtheta(fixedthetaSEXP);
-    rcpp_result_gen = Rcpp::wrap(fixdynamic(uld, theta, fixedtheta));
+    Rcpp::traits::input_parameter< Eigen::Matrix<int, Eigen::Dynamic, 1> >::type isfixed(isfixedSEXP);
+    rcpp_result_gen = Rcpp::wrap(fixdynamic(pfun, theta, isfixed));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -90,6 +101,31 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< pADFun& >::type uld(uldSEXP);
     Rcpp::traits::input_parameter< transform<a1type>& >::type tran(tranSEXP);
     rcpp_result_gen = Rcpp::wrap(reembed(uld, tran));
+    return rcpp_result_gen;
+END_RCPP
+}
+// fixindependent
+pADFun fixindependent(pADFun& pfun, veca1 x, Eigen::Matrix<int, Eigen::Dynamic, 1> isfixed);
+RcppExport SEXP _scorematchingad_fixindependent(SEXP pfunSEXP, SEXP xSEXP, SEXP isfixedSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< pADFun& >::type pfun(pfunSEXP);
+    Rcpp::traits::input_parameter< veca1 >::type x(xSEXP);
+    Rcpp::traits::input_parameter< Eigen::Matrix<int, Eigen::Dynamic, 1> >::type isfixed(isfixedSEXP);
+    rcpp_result_gen = Rcpp::wrap(fixindependent(pfun, x, isfixed));
+    return rcpp_result_gen;
+END_RCPP
+}
+// keeprange
+pADFun keeprange(pADFun& pfun, Eigen::Matrix<int, Eigen::Dynamic, 1> keep);
+RcppExport SEXP _scorematchingad_keeprange(SEXP pfunSEXP, SEXP keepSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< pADFun& >::type pfun(pfunSEXP);
+    Rcpp::traits::input_parameter< Eigen::Matrix<int, Eigen::Dynamic, 1> >::type keep(keepSEXP);
+    rcpp_result_gen = Rcpp::wrap(keeprange(pfun, keep));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -148,19 +184,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// ptapell2
-pADFun ptapell2(veca1 z_ad, veca1 theta_ad, Rcpp::XPtr<llPtr> llfXPtr, transform_a1type& tran, Eigen::Matrix<int, Eigen::Dynamic, 1> fixedtheta, bool verbose);
-RcppExport SEXP _scorematchingad_ptapell2(SEXP z_adSEXP, SEXP theta_adSEXP, SEXP llfXPtrSEXP, SEXP tranSEXP, SEXP fixedthetaSEXP, SEXP verboseSEXP) {
+// tape_bdryw_inbuilt
+pADFun tape_bdryw_inbuilt(std::string name, veca1 x, const double& acut);
+RcppExport SEXP _scorematchingad_tape_bdryw_inbuilt(SEXP nameSEXP, SEXP xSEXP, SEXP acutSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< veca1 >::type z_ad(z_adSEXP);
-    Rcpp::traits::input_parameter< veca1 >::type theta_ad(theta_adSEXP);
-    Rcpp::traits::input_parameter< Rcpp::XPtr<llPtr> >::type llfXPtr(llfXPtrSEXP);
-    Rcpp::traits::input_parameter< transform_a1type& >::type tran(tranSEXP);
-    Rcpp::traits::input_parameter< Eigen::Matrix<int, Eigen::Dynamic, 1> >::type fixedtheta(fixedthetaSEXP);
-    Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
-    rcpp_result_gen = Rcpp::wrap(ptapell2(z_ad, theta_ad, llfXPtr, tran, fixedtheta, verbose));
+    Rcpp::traits::input_parameter< std::string >::type name(nameSEXP);
+    Rcpp::traits::input_parameter< veca1 >::type x(xSEXP);
+    Rcpp::traits::input_parameter< const double& >::type acut(acutSEXP);
+    rcpp_result_gen = Rcpp::wrap(tape_bdryw_inbuilt(name, x, acut));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -189,18 +222,17 @@ BEGIN_RCPP
 END_RCPP
 }
 // tapesmd
-pADFun tapesmd(pADFun& uldtape, transform<a1type>& tran, manifold<a1type>& M, std::string weightname, const double& acut, bool verbose);
-RcppExport SEXP _scorematchingad_tapesmd(SEXP uldtapeSEXP, SEXP tranSEXP, SEXP MSEXP, SEXP weightnameSEXP, SEXP acutSEXP, SEXP verboseSEXP) {
+pADFun tapesmd(pADFun& uldtape, transform<a1type>& tran, manifold<a1type>& M, pADFun& bdrywtape, bool verbose);
+RcppExport SEXP _scorematchingad_tapesmd(SEXP uldtapeSEXP, SEXP tranSEXP, SEXP MSEXP, SEXP bdrywtapeSEXP, SEXP verboseSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< pADFun& >::type uldtape(uldtapeSEXP);
     Rcpp::traits::input_parameter< transform<a1type>& >::type tran(tranSEXP);
     Rcpp::traits::input_parameter< manifold<a1type>& >::type M(MSEXP);
-    Rcpp::traits::input_parameter< std::string >::type weightname(weightnameSEXP);
-    Rcpp::traits::input_parameter< const double& >::type acut(acutSEXP);
+    Rcpp::traits::input_parameter< pADFun& >::type bdrywtape(bdrywtapeSEXP);
     Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
-    rcpp_result_gen = Rcpp::wrap(tapesmd(uldtape, tran, M, weightname, acut, verbose));
+    rcpp_result_gen = Rcpp::wrap(tapesmd(uldtape, tran, M, bdrywtape, verbose));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -212,19 +244,22 @@ static const R_CallMethodDef CallEntries[] = {
     {"_scorematchingad_abort_recording", (DL_FUNC) &_scorematchingad_abort_recording, 0},
     {"_scorematchingad_taylorApprox_currentdynparam", (DL_FUNC) &_scorematchingad_taylorApprox_currentdynparam, 4},
     {"_scorematchingad_taylorApprox", (DL_FUNC) &_scorematchingad_taylorApprox, 5},
+    {"_scorematchingad_avgrange", (DL_FUNC) &_scorematchingad_avgrange, 1},
     {"_scorematchingad_set_cppad_error_handler", (DL_FUNC) &_scorematchingad_set_cppad_error_handler, 0},
     {"_scorematchingad_test_Rcpphandler", (DL_FUNC) &_scorematchingad_test_Rcpphandler, 0},
     {"_scorematchingad_fixdynamic", (DL_FUNC) &_scorematchingad_fixdynamic, 3},
     {"_scorematchingad_reembed", (DL_FUNC) &_scorematchingad_reembed, 2},
+    {"_scorematchingad_fixindependent", (DL_FUNC) &_scorematchingad_fixindependent, 3},
+    {"_scorematchingad_keeprange", (DL_FUNC) &_scorematchingad_keeprange, 2},
     {"_scorematchingad_tape_Jacobian", (DL_FUNC) &_scorematchingad_tape_Jacobian, 1},
     {"_scorematchingad_tape_Hessian", (DL_FUNC) &_scorematchingad_tape_Hessian, 1},
     {"_scorematchingad_tape_gradoffset", (DL_FUNC) &_scorematchingad_tape_gradoffset, 1},
     {"_scorematchingad_tape_logJacdet", (DL_FUNC) &_scorematchingad_tape_logJacdet, 1},
     {"_scorematchingad_tape_swap", (DL_FUNC) &_scorematchingad_tape_swap, 1},
-    {"_scorematchingad_ptapell2", (DL_FUNC) &_scorematchingad_ptapell2, 6},
+    {"_scorematchingad_tape_bdryw_inbuilt", (DL_FUNC) &_scorematchingad_tape_bdryw_inbuilt, 3},
     {"_scorematchingad_getllptr", (DL_FUNC) &_scorematchingad_getllptr, 1},
     {"_scorematchingad_tape_uld_inbuilt", (DL_FUNC) &_scorematchingad_tape_uld_inbuilt, 3},
-    {"_scorematchingad_tapesmd", (DL_FUNC) &_scorematchingad_tapesmd, 6},
+    {"_scorematchingad_tapesmd", (DL_FUNC) &_scorematchingad_tapesmd, 5},
     {"_rcpp_module_boot_manifolds", (DL_FUNC) &_rcpp_module_boot_manifolds, 0},
     {"_rcpp_module_boot_ADFun", (DL_FUNC) &_rcpp_module_boot_ADFun, 0},
     {NULL, NULL, 0}